reservation complete

See merge request sambhavk/java-training-flight-reservation-system!7

+++ src/main/java/org/ideyalabs/flights/dto/FlightResponseDto.java
@@ -10,7 +10,7 @@ import java.time.LocalDateTime;
 @AllArgsConstructor
 @NoArgsConstructor
 public class FlightResponseDto {
-    private Integer flightId;
+    private Long flightId;
     private String flightName;
     private LocalDateTime DepartureTime;
     private LocalDateTime ArrivalTime;
